mysite/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Cython==3.0.0
Django==4.2.3
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
docopt==0.6.2
idna==3.4
mysql-connector-python==8.1.0
Pillow==10.0.0
pipreqs==0.4.13
protobuf==4.21.12
requests==2.31.0
sqlparse==0.4.4
telnetlib3==2.0.4
typing_extensions==4.7.1
urllib3==2.0.4
yarg==0.1.9
